introduce sequence id so packages get not dropped as duplicate
nRF24L01+ Datasheet:
"When several data packets are lost on the link, the PID fields may
become equal to the last received PID. If a packet has the same PID as
the previous packet, nRF24L01+ compares the CRC sums from both packets.
If the CRC sums are also equal, the last received packet is considered
a copy of the previously received packet and discarded"

This causes problems when load is always the same in a tree so
introduce a sequence id to avoid "loosing" packages!